+++ bugsnag-android-core/src/androidTest/java/com/bugsnag/android/ThreadStateTest.kt
@@ -21,7 +21,7 @@ class ThreadStateTest {
         null,
         true,
         1000,
-        100,
+        1000,
         ThreadSendPolicy.ALWAYS,
         Collections.emptyList(),
         NoopLogger,
@@ -65,7 +65,7 @@ class ThreadStateTest {
             trace,
             true,
             1000,
-            100,
+            1000,
             ThreadSendPolicy.ALWAYS,
             Collections.emptyList(),
             NoopLogger,
@@ -92,7 +92,7 @@ class ThreadStateTest {
             trace,
             true,
             1000,
-            100,
+            1000,
             ThreadSendPolicy.ALWAYS,
             Collections.emptyList(),
             NoopLogger,
@@ -117,7 +117,7 @@ class ThreadStateTest {
             trace,
             true,
             1000,
-            100,
+            1000,
             ThreadSendPolicy.ALWAYS,
             Collections.emptyList(),
             NoopLogger,
@@ -163,7 +163,7 @@ class ThreadStateTest {
             exc,
             true,
             1000,
-            100,
+            1000,
             ThreadSendPolicy.ALWAYS,
             Collections.emptyList(),
             NoopLogger,
@@ -201,7 +201,7 @@ class ThreadStateTest {
             trace,
             true,
             2,
-            100,
+            1000,
             ThreadSendPolicy.ALWAYS,
             Collections.emptyList(),
             NoopLogger,
@@ -230,7 +230,7 @@ class ThreadStateTest {
             exc,
             true,
             4,
-            100,
+            1000,
             ThreadSendPolicy.ALWAYS,
             Collections.emptyList(),
             NoopLogger,
@@ -258,7 +258,7 @@ class ThreadStateTest {
             trace,
             true,
             1000,
-            100,
+            1000,
             ThreadSendPolicy.NEVER,
             Collections.emptyList(),
             NoopLogger,
@@ -268,6 +268,26 @@ class ThreadStateTest {
         val json = streamableToJsonArray(state)
 
         assertEquals(0, json.length())
+    }
+
+    @Test
+    fun testThreadCollectionTimeLimitMillis() {
+        val currentThread = Thread.currentThread()
+        val allThreads = allThreads()
+        val state = ThreadState(
+            trace,
+            true,
+            1000,
+            0,
+            ThreadSendPolicy.ALWAYS,
+            Collections.emptyList(),
+            NoopLogger,
+            currentThread,
+            allThreads
+        )
+        val json = streamableToJsonArray(state)
+
+        assertEquals(1, json.length())
     }
 
     private fun verifyCurrentThreadStructure(